+++ girder_annotation/girder_large_image_annotation/rest/annotation.py
@@ -661,19 +661,29 @@ class AnnotationResource(Resource):
         .param('sources', 'An optional comma separated list that can contain '
                'folder, item, annotation, annotationelement, datafile.',
                required=False)
+        .jsonParam(
+            'compute', 'A dictionary with keys "columns": a list of columns '
+            'to include in the computation; if unspecified or an empty list, '
+            'no computation is done, "function": a string with the name of '
+            'the function, such as umap, "params": additional parameters to '
+            'pass to the function.  If none of the requiredKeys are '
+            'compute.(x|y|z), the computation will not be performed.  Only '
+            'rows which have all selected columns present will be included in '
+            'the computation.',
+            paramType='formData', requireObject=True, required=False)
         .errorResponse('ID was invalid.')
         .errorResponse('Read access was denied for the item.', 403),
     )
     @access.public(cookie=True, scope=TokenScope.DATA_READ)
     def getItemPlottableData(
-            self, item, keys, adjacentItems, annotations, requiredKeys, sources=None):
+            self, item, keys, adjacentItems, annotations, requiredKeys, sources=None, compute=None):
         user = self.getCurrentUser()
         if adjacentItems != '__all__':
             adjacentItems = str(adjacentItems).lower() == 'true'
         sources = sources or None
         data = utils.PlottableItemData(
             user, item, annotations=annotations, adjacentItems=adjacentItems,
-            sources=sources)
+            sources=sources, compute=compute)
         return data.data(keys, requiredKeys)
 
     def getFolderAnnotations(self, id, recurse, user, limit=False, offset=False, sort=False,