+++ girder_annotation/test_annotation/test_annotations_rest.py
@@ -866,7 +866,7 @@ class TestLargeImageAnnotationElementGroups:
             },
         )
         assert utilities.respStatus(resp) == 200
-        assert len(resp.json) == 2
+        assert len(resp.json) == 5
 
         resp = server.request(
             path=f'/annotation/item/{itemSrc["_id"]}/plot/list',